add operator-managed resources to image-loader
